Move jest-runner-vscode config to correct location
